+++ init.go
@@ -1,7 +1,7 @@
 package rui
 
 func init() {
-	//resources.init()
-	defaultTheme.init()
-	defaultTheme.addText(defaultThemeText)
+	if theme, ok := CreateThemeFromText(defaultThemeText); ok {
+		defaultTheme = theme
+	}
 }