+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added "user-data" property
 * Added "focusable" property
 * Added ReloadTableViewData, AllImageResources functions
+* Added Theme interface, NewTheme, CreateThemeFromText, and AddTheme functions
 
 # v0.5.0